refactor(NcAppNavigationList): migrate to Typescript
