+++ src/components/index.ts
@@ -22,7 +22,7 @@ export { default as NcAppNavigation } from './NcAppNavigation/index.ts'
 export { default as NcAppNavigationCaption } from './NcAppNavigationCaption/index.js'
 export { default as NcAppNavigationIconBullet } from './NcAppNavigationIconBullet/index.js'
 export { default as NcAppNavigationItem } from './NcAppNavigationItem/index.js'
-export { default as NcAppNavigationList } from './NcAppNavigationList/index.js'
+export { default as NcAppNavigationList } from './NcAppNavigationList/index.ts'
 export { default as NcAppNavigationNew } from './NcAppNavigationNew/index.js'
 export { default as NcAppNavigationNewItem } from './NcAppNavigationNewItem/index.js'
 export { default as NcAppNavigationSearch } from './NcAppNavigationSearch/index.js'
